small fix to submissions, previously was breaking due to querying non-existant submissions for the assignment

+++ LMSV1/Pages/Courses/Assignments/Submissions/Submit.cshtml.cs
@@ -122,11 +122,12 @@ namespace LMSV1.Pages.Courses.Assignments.Submissions
                 Submission submission = _context.Submissions
                 .AsNoTracking()
                 .FirstOrDefault(a => a.UserID == UserID && a.AssignmentID == Assignmentid);
+                if (submission != null)
+                {
+                    _context.Submissions.Remove(submission);
 
-                _context.Submissions.Remove(submission);
-
-                _context.SaveChanges();
-
+                    _context.SaveChanges();
+                }
             } while (submission != null);
         }
     }
